test(fixture): move fixture file
